+++ src/game/scene.rs
@@ -5,7 +5,7 @@ use super::paddle::DEFAULT_PADDLE_WIDTH;
 pub const SCREEN_WIDTH: u16 = 800;
 pub const SCREEN_HEIGHT: u16 = 400;
 
-pub const PADDLE_MARGIN: u8 = DEFAULT_PADDLE_WIDTH;
+pub const PADDLE_MARGIN: u16 = DEFAULT_PADDLE_WIDTH;
 
 pub enum Edges {
     Top,
